requirements.txt
astroid==2.2.5
boto3==1.11.4
botocore==1.14.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.14
dataclasses
docutils==0.15.2
graphviz==0.8.4
holidays==0.9.12
hupper==1.9.1
idna==2.6
isort==4.3.21
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lunardate==0.2.0
matplotlib==3.1.2
mccabe==0.6.1
missingno==0.4.2
mxnet==1.4.1
numpy
pandas==0.25.3
PasteDeploy==2.0.1
patsy==0.5.1
plaster==1.0
plaster-pastedeploy==0.7
pmdarima==1.5.2
pydantic==1.3
pylint==2.3.1
PyMeeus==0.3.6
pyparsing==2.4.6
pyramid==1.10.4
pystan==2.19.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.20.0
s3transfer==0.3.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
setuptools-git==1.2
six==1.14.0
sklearn==0.0
statsmodels==0.10.2
tqdm==4.41.1
translationstring==1.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.24.2
venusian==3.0.0
WebOb==1.8.5
wrapt==1.11.2
zope.deprecation==4.4.0
zope.interface==4.7.1
gluonts
fbprophet